Prevent multiple event deletion when event page is deleted
The page delete handler did not check whether the event was already
deleted. Thus, it was possible for the event deletion to happen more
than once, if the page was also deleted more than once (while being
restored or recreated in the meantime). This was mostly just a no-op and
overall useless, but it had the unwanted side effect of updating the
deletion timestamp, hence losing information on when the deletion
originally occurred.

So, return early if the associated registration is already deleted. Add
a test to verify the change.

Bug: T395351
Change-Id: Iaea5983742cfdae09b3c908d4c47634953793b4d