Fold EventTypesFormatter into EventTypesRegistry
The registry on its own was mostly only used to support the formatter,
where the actual data lived. As we discussed in an engineering meeting,
having separate constants for each type makes it a bit harder to keep
everything up-to-date, because there are multiple lists that need to be
kept in sync. Currently, we also don't really have any reasons to access
the individual types (e.g., via constants) to special-case them
anywhere.

So, merge the formatter and registry into a single class that provides a
list of grouped event types, as well as i18n. This is similar to other
existing registries, like article topics.

This patch copies the code almost verbatim. It will then be changed in a
subsequent patch to consolidate all the information about event types.

Change-Id: Ic3064c89bd5989f29872b8bd15391d278a8f40c5