Re-order event properties
This list exists in multiple places as method parameters, API payloads,
form fields, messages, test data, etc. Sadly, due to the addition of
multiple features over time, each list was expanded semi-independently
from the others, with their orderings eventually drifting away more and
more. Aside from the obvious inconsistency, this makes it more difficult
to verify the correctness of any individual list, or to see if a given
property is already present/implemented. It's also more mental effort
because there isn't a single ordering to memorize that works everywhere.

Currently, the only ordering that most people are exposed to, and whose
correctness/stability is therefore important, is the ordering used for
form fields in the registration form.

So, update all the existing lists (except for database stuff) to follow
the same ordering as form fields. Leave additional metadata (timestamps)
at the end.

This refactoring is mostly a no-op without visible effects, except for
the event types form field that gets moved up the form according to the
AC.

See also https://www.mediawiki.org/wiki/Special:Diff/7659517 for the API
documentation update.

Bug: T386273
Change-Id: I4612f7f2982d39a95cd73c1a41eaee11a21d14c5