Avoid using WikiConfig for checking event type feature flag
The $wikiConfig property here holds a reference to the subset of
configuration that we allow to be customized on-wiki (via
CommunityConfiguration); it is not a general reference to the
MainConfig.

The CampaignEventsEnableEventTypes feature flag is not
community-overridable, so it should be read from MainConfig directly,
rather than relying on the CC provider falling back to it.

Change-Id: I935d1fe4d37a14c2bef495056e08977de5e9ef94