Remove l10n from EventTypesRegistry
Leave i18n only, but have callers be responsible for i18n, as they
typically have more context available (e.g. implementors of
MessageLocalizer). Accordingly, drop the MessageFormatterFactory
parameter from the constructor.

Also update the i18n method to accept a list of multiple types, as
that's going to be the most common (or only) scenario in practice. This
is also modelled after the existing pattern in
ArticleTopicFiltersRegistry and WikiLookup.

Change-Id: Id7656e23e2ab47fab909ca9f6165444f954b1585