SpecialEventDetails: hide stats tab if there are no participants
We cannot report ranges when we don't know the total number of
participants; in practice, it would result in a division by zero.

So, have the caller check the number of participants, and then pass that
in.

Note that statistics can still become meaningless as more participants
unregister (T349650); this patch only addresses the extreme case where
*all* participants unregister.

Bug: T387243
Change-Id: If4e404080d3b606ca0516f0d066772f1d8501f16