refactor: Use MediaWikiConfigRouter for mixed configuration
MediaWikiConfigRouter is designed to pick between configuration
intentionally only available in PHP Settings and configuration that
has been added to CommunityConfiguration.

MediaWikiConfigReader, on the other hand, should only ever be used for
configuration found in CommunityConfiguration. For now, it is still
falling back to PHP Settings, but that is expected to be replaced with
throwing an exception. (Maybe we should also rename this service/class?)
Also, it is creating an INFO-level log entry in logstash, which is
intended to help us find instances like this that should be reworked.

Bug: T387452
Change-Id: If71d4fa953accfb9143bcd301689eaab0d7c325b