i18n: Rename "location" to "participation options"
"Location" is non-standard terminology that we used to refer to "event
meeting type" in some places. We are now consistently using
"participation options" everywhere.

This patch only updates message contents and renames some messages. The
qqq for some of the renamed messages will be updated in a subsequent
patch, after the next translations export, to make sure that the rename
is successful.

The code itself (like variable names) will also be updated in a
subsequent patch.

Bug: T387493
Change-Id: Id8a2d017918bb9ba126498f55b52f7f929ec5964