Add new country dropdown field
When writing to the new country schema, display a country dropdown that
gives us a country code, instead of a free-text value. AddressStore will
add in an interim free-text country (in English) if required to also
write to the old schema.

For the WRITE_OLD case, add an assertion to make sure we don't lose
data. This will never happen in practice, unless we roll back the
migration (and in that case, an exception is probably fine given the
emergency).

For the WRITE_NEW case, note that we are discarding the free-text value,
if the organizer edits the event before the migration script updates it.
However, since the country code is a required value, the organizer will
still be required to fill it out, and be notified of the change. This is
not ideal, but it seemed preferable to the other possible approaches
(see also T397476#10980178).

Bug: T397271
Change-Id: Ic1c8bdb2c914a5bd61916bbb90c019765ef74c67