Introduce a value object to represent addresses
This finally lets us make EventStore unaware of the underlying schema.
For now, Address objects provide the "address without country"
separately from the country. This will be changed in a later patch, as
callers should need either the full address (including country), or the
country.
A subsequent patch will also expand usage of the Address object to other
classes (like EventRegistration), so that callers will be easier to
migrate from country to country code. The Address class will also be
used as a temporary compatibility layer between the two formats.

Bug: T397273
Bug: T397867
Bug: T317578
Change-Id: Iae18396025073c205a4f6ce3cb8a44d0e37082a1