Use null for event address instead of empty string
Previously, we would have an empty string as meeting address when only
the country (but no address) was specified. This was mostly harmless
because we typically check for truthiness when deciding whether to
display the address. However, the event page header is an exception, in
that it checks `$address !== null`. There, a missing address
accidentally resulted in the mapPin icon being shown without any text.

So, convert empty strings to null when creating the EventRegistration
object.

Change-Id: I86d3ddca76259980affeab510ee10a346196fbf0