Include country in the address in the registration header
Previously, the registration header would display the address for
in-person events, but not the country. This was a weird choice, and in
contrast with every other place where we display the address (for
example, the "more details" modal, Special:EventDetails, and the
confirmation email).

This code was introduced in Ief09dc778ade8e (June 2022), and already in
the first revision it was showing the address only, and no country. The
same was true for the "more details" dialog. The bug was then identified
in QA (T305699#8031350). It was then fixed shortly after in
If048c1f8b8235d54 for the dialog, but not for the header! It's pretty
obvious that this wasn't intentional. So, include the country in the
registration header like everywhere else.

Note that:
- Newlines are not preserved (there's no special CSS), but that's
  already true for the address, which can include newlines on its own.
  That is unchanged with this patch.
- The address gets truncated to 30 characters, so the country may not be
  shown anyway.

Bug: T397867
Change-Id: Id0de80b1c209ed4a7a5f6b55e7be1bfa136d108b