Have AddressStore::acquireAddressId() take Address
The implementation will need to change for the country code migration,
and it won't rely on the "address + \n + country" trick. So, have the
method take an Address object and build the query on its own.

Also add an assertion to the Address constructor that  address and
country must not be both null. This is already the case, and some places
make this assumption when they check that they got an Address object
(without actually checking what's inside).

Fix tests that were breaking the assumption in question, and expand
tests for acquireAddressId.

Bug: T397273
Change-Id: I0a7bfca9e4067acb53742c5485dda45d6405e63b