Start reading country schema migration flag in AddressStore
Update code paths that will depend on the migration stage to check what
stage we're in. Move the existing code to the appropriate *_OLD branch.
*_NEW branches will be written in subsequent patches.

Add tests for all possible combinations of migration stage and stored
data. For now, we only run tests for MIGRATION_OLD, as everything else
still needs to be implemented.

Bug: T397544
Bug: T397273
Change-Id: Ib17716c8d65aa268333d82a8a4bbefcec049e90b