Rename EventRegistration::$meetingAddress to $address for cache compat
In I3a262b775754919aa, the `EventRegistration::$meetingAddress` property
was changed from `string|null` to `Address`. This is a
backwards-incompatible change for serialized EventRegistration objects
stored in cache, so I bumped the cache version in
`EventStore::getEventByPage` at the same time.

However, the cache version is stored as part of the cached value, and it
is only read and compared *after* the value has been retrieved from
cache; and, in the case of serialized objects, unserialized. This means
that we would fail with a TypeError even before checking the version.

So, make the property use a new name. This way, we avoid the TypeError
and will later discard the stored object.

A subsequent patch will also refactor the test to catch these issues.

Bug: T398413
Change-Id: I82c89438106a3bf9267be535edbb722cc5d60ed2