Add country code to Address and use it for formatting
Add a third parameter for the country code and an associated getter. The
parameter is currently nullable, but it will become non-nullable once
the migration is over.

Expand the test suite in AddressStoreTest to include a lot more
scenarios. Update other tests to pass a valid country name and, in
some cases (where supported), a country code. Add a parameter to test
for the newly-introduced exceptions.

For now, no actual code is passing a non-null value, only tests. Also,
the value is unused outside of EventFormatter. AddressStore requires
that a country code be passed iff the migration stage has WRITE_NEW.
This will change in subsequent patches, as the code for `*_NEW` stages
is written.

In EventFormatter, try to use the country code when available, and fall
back to the free-text country otherwise. Add a method to CountryProvider
to get the translation for a single country code and language.

Bug: T397867
Change-Id: I39f4edcd4e0c6540c95637f749efee7a90f51737