Add EventFormatter::formatAddress, drop Address::toString
Adding this to Address was my mistake. Soon, formatting an address will
require a language code and a service to get a localised country name in
that language for a given country code. We can't have this state inside
the Address value object in the long term.

So, move `Address::toString()` to `EventFormatter::formatAddress()` and
migrate callers accordingly, injecting an EventFormatter where needed.

Note that the language code is currently unused, but it's being added
here so we won't have to update callsites again once we start using it.

Bug: T397867
Change-Id: I2ae712e7735c0c548a69ca37916bcc0938302dc6