Add event country code to EventFactory
For the time being, the parameter is nullable and can be used alongside
the free-text country. If a code is provided, then we check for its
validity. Additionally, the code is required if we need to write to the
new schema; the error message for this is hardcoded in English as it'll
be temporary, and in practice it should only be seen by API users.

All non-test callers are still passing null, as they'll be updated in
other patches/tasks.

Update tests and add test cases for invalid country code.

Bug: T397867
Change-Id: Ieaa647a649a6e4c6e763005a539250d8282c5961