Handle empty result sets in EventStore::newEventsFromDBRows
The method would previously throw an exception because of empty WHERE
clauses passed to Database methods. The API endpoints that rely on the
method (list events by participant and by organizer) would currently
throw in production when given a user without events.

So, return early if there are no events, without querying additional
tables.

Add a regression test for both methods that rely on this code. This is
done with data providers so that the tests could later be expanded if we
add new features to these methods (e.g., I2e3f37cd3f304c).

Change-Id: I89cb59210635a4d73569a93bb069b75a22f34dac