Tidy up EventPagerTrait::getSubqueryInfo
Inline arrays inside the queryinfo return value, as is usually done.
Extract the list of campaign_event fields into its own variable, and use
it for both the field list and the GROUP BY clause. Add a comment
documenting why we need to GROUP BY everything.

Omit the list of fields from the outer query, and rely on the subquery
providing useful aliases.

Change-Id: If1513dcc595efdc045a94ce9130a2c4620259f6f