Make registration form enhancements more robust for browser tests
TimeFieldsEnhancer modifies the time fields in a variety of ways that
can occasionally cause issues, especially if the form is submitted too
quickly.

Making the logic more lenient isn't easy because there'd be several
places to update; especially given the double validation in HTMLForm and
EventFactory.

So, leave the code as-is but make selenium wait until the field have been
enhanced. Also clean up the code a bit:
- Rename `$formRoot` to just `$root`, as the element passed in is
  actually the entire document (`$formRoot` implied that it was just the
  form).
- Introduce local variables to reduce duplication and improve
  readability.
- Skip date checks for aggregated answers when enabling registration,
  and add a comment explaining why we aren't checking for existing
  answers in one of the branches.

Apply the same wait-for-enhancement logic to the organizer selector.

Bug: T393428
Change-Id: Ieb6407406ca3942006ed8eaf8612b4fd01b081d0