resources: Fix promise rejection handlers
In Idac8920aeabbec we switched from .done()/.fail() to .then(). A key
difference is that the latter doesn't return the original promise;
instead, it returns a different promise based on the success/failure
handlers, and that promise will be successful if the failure handler
does not return anything.

So, update the affected failure handlers to throw the received error,
when the intention is to let other parts of the code react to the
failure.

Bug: T401873
Change-Id: I9b3ec3e34443bd9e4df846a97529804b381668a1