Drop duplicated rows in UpdateCountriesColumn
Once all countries have been updated, scan the table to see if we got
any duplicates, and if so, keep only one row for each, updating
references in ce_event_address.

The script attempts to print the number of duplicates, but this is
largely inaccurate when in dry-run mode, because without updating the
rows we can't tell if two of them will get mapped to the same value.

Mention in AddressStore that the key used for acquiring an address ID
must always be unique, regardless of what it is exactly.

Refactor the script a bit so that each step now has its own method, for
readability.

Update tests to use different addresses (so we don't deduplicate all of
them), and add a new row for testing deduplication.

Note that the script won't correctly fixup existing events that already
have two addresses due to the bug described in the task: instead, it
assumes that events have a single address and that a single UPDATE for
ce_event_address records is sufficient.

Bug: T402239
Change-Id: I069835ce83a0e50f4051e6c6aad6d1f2a11942e0