Avoid PHP notice in AbstractEventRegistrationSpecialPage (country field)
The 'EventMeetingCountry' field is no longer present when the migration
stage includes WRITE_NEW, so we can't access it unconditionally.

Remove it from $nullableFields, and adjust its value separately. This
way it'll also be easier to drop this code when the field itself is
dropped.

Copied from I35202674da10e621e.

Bug: T402441
Change-Id: I85ba2bfdca29de3088cd57213b162d2c68e1c1a9