Add edit association dialog
For source edits and VE/MF page creations, add a handler for the
BeforePageDisplay hook that obtains a list of valid user events and, if
the list is not empty, adds the JS module to show the dialog.

For VE edits (but not page creations), register the module as a VE
plugin, so that it's loaded iff VE is loaded.

The list of events has an arbitrary limit of 50 entries, which should be
much more than enough for real-world usage.

The dialog itself uses Codex (in Vue), and it's parly based on
I6b11d5887eca.

Also add the needed dependencies and scaffolding for unit testing in
Jest, with the boilerplate mostly copied from other extensions. Add unit
tests for the Vue application.

Bug: T400953
Change-Id: I9bd8c0e65a0e895f404fb405f3cea59a849bd0ea