EventFactory: implement contribution tracking validation
When the event is configured to track contributions, verify that it's
either online, or not in a disallowed country. The country list is
configurable, and empty by default. In Wikimedia production, it will be
populated with data from https://w.wiki/FBUa in a separate config
change. Then verify that the event type is contribution-related;
introduce a new getter in EventTypesRegistry to validate this.

Despite validation being implemented, all callers under user control
(API and special page) are still only passing false. These will be
updated in the next patch.

Bug: T402353
Change-Id: I7ac531c1889fc67d0b9d7f8366f99b01feeea73f