Fix minor signature mismatch in EventStore methods
The $limit parameter is not nullable in the interface, only in the
inherited methods. I'm not sure if this was intentional, since a limit
should always be provided as a safety net, and all callers are already
passing one. So, make the limit not nullable in EventStore, and simplify
the method body accordingly.

Change-Id: I2d8f22648c88d342c7a851b38e1c059e04d6f52a