Read event_track_contributions in the event pagers
The DB row is used to construct EventRegistration objects, and therefore
it needs to have all the fields set. This code was working fine with the
CampaignEventsEnableContributionTracking set to false, but not when it
was set to true, as the event_track_contributions field was missing. As
a result, Special:AllEvents was throwing an uncaught exception. So, add
the field conditionally based on the feature flag.

Follow-up to I431fdf53103834c051.

Change-Id: I0995984232e9f401a48a97446de676911ace293f