Use extension registry to require 'cldr'
Due to T398224, extension registry could not find the cldr extension
cause it was named "CLDR". I7139c55e58fc worked around the case
mismatch.

In I38f7df13e787 the name in the extension was made all lower case to
match the repository name and we can now "requires" it.

Bug: T398224
Depends-On: I38f7df13e787c5aa0277d5a204a4b176316df46c
Change-Id: Ib4462865ab9af54b36c7a4874995c3af7983f444