Fix disallowed country codes format for EventFactory
The config variable is a map of country codes to country names, but
EventFactory wants a list of country codes (the names are for
readability and comparison only). So, add a missing `array_key` to
convert between the two.

Note that the unit test was already passing the expected format; and
phan cannot warn about this because it doesn't know the type of the
config setting.

Bug: T402353
Change-Id: I17d9932be3c149199f5c71ead11d5c9400fa05d0