Rename PageEventIngress to EventPageEventIngress
Soon we will add a separate event ingress for page-related events, but
used to update edit associations. So, rename the existing event ingress
to clarify that it's only for event pages.

Also fix indentation in a couple places, as well as a test that wasn't
very useful because of testing the return value of a void method.

Bug: T403200
Change-Id: I793eb405a76edf13b0ebb151194fcd99220b2147