EventContribution: handle page moves and page/revision deletes
Introduce a new EventIngress for page-related events that listens to
page moves, deletions, creations (for undeletions only) and revision
visibility changes. If the page affected is determined to have at least
one edit associated with an event, then we schedule a job to update the
relevant records in the contributions storage. Update
EventContributionStore with all the methods needed to perform such
updates, using batching to avoid potentially writing a lot of rows at
once.

Bug: T403200
Change-Id: Id7a92f1069d0de1526000556b27f734dd8a9db04