EventContribution: properly format foreign titles
When computing a new contribution, check if the page is local. If it
isn't, rather than using TitleFormatter (which would not produce the
correct result), make a cross-wiki API request to get the formatted
title. Error handling is strict, assuming that there will never be any
failures. This should mostly be the case, and we can iterate if needed.
The result is cached for a week, so that subsequent edits to the same
page will not require a new API request.

Add some semi-unit tests for this, covering both the local and foreign
cases. Also add a database integration test to cover the entire metrics
computation.

In addition, and in part to make tests pass:
- Use WikiAwareEntity::LOCAL when the given wiki ID matches the current
  wiki's ID. Some core services already do this. It makes sure that the
  wiki assertions in page do not fail if the PageIdentity object uses
  ::LOCAL but is compared against the full ID.
- Do not try to get the parent revision when there isn't one. It's just
  a pointless query looking for rev_id=0.
- Structure EventContributionComputeMetricsTest a bit more similarly to
  other unit tests, without holding class state that might get
  out-of-date.

Bug: T404899
Depends-On: Icf6a008433fed5d1a165d5f36fc9a61626bfd375
Change-Id: If032b592c0237b8bed811203c9ba77ada91e0fee