Update database schema to store usernames associated with contributions
This will allow us to sort the table by username. This patch modifies
the database only, as well as the bare minimum of queries required not
to break tests. Notably, the updated query in EventContributionStore
uses the empty string as a placeholder. This will be changed in a
subsequent patch.

Note that in `assertValidRow`, `isset` needs to be changed to
`property_exist` to support the column being nullable. This is why it
was done this way in EventStore.

NOTE: the change here is done in a backwards-INcompatible way, the
reason being that this code is disabled everywhere except for local test
environments, and the new table was introduced recently (T400719) and is
not part of any release. Hence, the schema change will need to be
applied manually in local environments. Before merging this patch, make
sure you can manually update the schema in beta, or coordinate with
someone who can do that immediately after the merge, to minimise
beta breakage.

Additionally, existing entries in local environments will not be
backfilled. It would probably be easiest to regenerate them, but if you
need to backfill them, you can use the script in
https://phabricator.wikimedia.org/P83943.

Bug: T404995
Change-Id: I8667675dc57516452eb46b18f7b95518c73b52bd