Add username to the event contribution entity layer
Add a new property/ctr parameter for the username, read it when storing
values, and write it when selecting stored data. Also set it to the
actual username when creating a new entry, updating the
EventContributionComputeMetrics class and related tests as needed.

Note that the value is still not used by the pager; this will be done
separately.

Bug: T404995
Change-Id: Ibe7dd23d0e5a5a56c84b52d5f32f10d64e8108bc