Check feature flag in ContributionAssociationPageEventIngress
This code was previously hitting the database and only then returning
due to lack of contribution data. However, this was not supposed to
happen as the feature is still experimental. The only reason why this
didn't cause a production error is that the ce_event_contributions table
has already been created even if it's unused. Nonetheless, this would
prevent us from easily applying the schema change for T404995, which is
being done under the assumption that no code is using the table.

Bug: T404995
Change-Id: I63d948af0c8d210879d6bacaeedb1003145ecde2