Make sure that every edit is associated with at most one event
Add a unique index to the DB schema to guarantee this at the DBMS level.
Then add a lookup method to EventContributionStore and call it from
EventContributionValidator to assert uniqueness. If the edit is already
associated with the given event, we just quietly treat it as success, if
not it's a user-visible error. Add tests for everything.

NOTE: the DB change here is done in a backwards-INcompatible way, the
reason being that this code is disabled everywhere except for local test
environments, and the new table was introduced recently (T400719) and is
not part of any release. Hence, the schema change will need to be
applied manually in local environments and beta.

Bug: T406146
Change-Id: I85b8ea4f602ee1d43c08671e779349f505ccec3c