SpecialEventDetails: hide contributions tab when tracking is disabled
Check whether the event has contirbution tracking enabled, and do not
show the tab if that isn't the case.

By doing this, we also no longer need the feature flag check, as it is
not possible to enable tracking while the feature flag is off. This way
we won't have to update this code again later.

Bug: T408005
Change-Id: I5a89d437858d86b836f3500a883cbceb0af72a9a