Make "Manage event" button on event page link to registration form
The "more details" button will soon become a direct link to
Special:EventDetails, so it wouldn't make sense for organizers to have a
duplicated link; this is already partly the case today, as
Special:EventDetails is already accessible by middle-clicking "More
details". So, make the "Manage event" button point to
Special:EditEventRegistration instead.

Bug: T403118
Change-Id: I9d626cf1cbf29e6f3a1ee9effbcd06d02860f031