EventContributionsPager: coalesce null names to string to fix sorting
The pagination logic uses comparisons such as `cec_user_name > 'Foo' OR
(cec_user_name = 'Foo' AND cec_timestamp > '123')` when going through
pages. This doesn't work well when the values in cec_user_name are null,
due to special null comparison semantics. So, introduce a stringified
version of the username to be used for sorting, using COALESCE to get
rid of nulls. In lack of an existing convention, this uses a double
underscore to make it clearer that it's not an actual column from the
table. Also note that there is no function in our RDBMS lib for
`COALESCE`, but thankfully this syntax is supported by all 3 DBMSs we
support.

Bug: T404995
Change-Id: I7c2fa17dd3648e39357453d51aa699104e90d37d