Clean up database schema for event country
Set the default migration stage to MIGRATION_NEW as the only value
currently supported. (The flag will be dropped in a separate patch.)

Drop the column for the free-text country, and make the country code not
nullable. Adjust the code so that it does not try to write or read
cea_country (this was wrong in the previous code, we shouldn't have done
that in the write_new/read_new branches).

Drop the migration script and its test, no longer useful.

Remove migration logic from AddressStoreTest so that it passes in the
updated database schema. The code itself (e.g., Address parameters) will
be cleaned up in a separate patch, and same for the migration code in
general.

Bug: T408931
Change-Id: Ia35492a69ba88a29b418d27028251f732319b33e