resources: Use Codex tokens in more places
Avoid hardcoding token values, not only is it bad practice, but it won't
work well under certain circumstances such as in dark mode.

All replacements are exact except for the answer statistics border. That
one is changed to the normal 'subtle' token, same as the border for <tr>
elements; not sure if the discrepancy was intentional in the first
place, but it looked a bit odd.

The TablePager's border is instead left untouched, as changing to the
'subtle' style there would be a more noticeable difference. That UI
might also be converted to CodexTablePager directly (T381349).

Bug: T409336
Change-Id: I77891a31d21ef51e0348698fc268e70519c633e5