Add option to hide contribution prompt to storage and behaviour layer
Add a parameter to RegisterParticipantCommand to choose between showing
and hiding the prompt. This takes class constants as values for better
call site interpretation/readability.

Make ParticipantsStore write the value (but not read it yet). This takes
a boolean as it's a low-level API and so probably fine (and avoid the
overhead of introducing more constants).

Higher-level callers (UI and API) are still passing a hardcoded default
for showing the prompt (status quo).

Bug: T408460
Change-Id: I9cdfec18a93637062742d0abeba160d243d315c3