Use JSON serialization for EventRegistration caching
This avoid PHP serialization, which is bad for multiple reasons, see
e.g.
https://www.mediawiki.org/wiki/Object_cache#Storing_structured_data.

Update the test accordingly, now we can be a bit more superficial, as
failures will also be caught by the serialization tests in
EventRegistrationTest. Use .enc rather than .json for the encoded blob
so that we don't need to exclude it from linters.

Bug: T410554
Change-Id: I7a713d154d01dfbe7bbe30acbcc5eee49623d264