Simplify return value of ParticipantsStore::addParticipantToEvent
Replace the MODIFIED_* bit flags with mutually exclusive constants, and
merge those for answers and visibility. This way we only need a single
conditional branch for partial updates, which makes the code easier to
reason about and to change without having to update a million places
(see for example PS 5 of I9cdfec18a9363706).

Change-Id: I298d955e1e3f83b8f631b28b508ead1efbea18dc