Check participant opt-out when listing events for association
Expand a bit the scope of
getEventsForContributionAssociationByParticipant so that it also checks
for the opt-out flag. This does mean that association might be possible
with events that are not included here (for which the user opted out but
all other conditions are met). However, this makes it easier to respect
the user preference without changing too many places, and also for
clients. So, document that this is the expected behaviour.

Bug: T408460
Change-Id: I50cf6ecab0ec2a3eaa7db7d0d0e13201ecd0dcf2