Update API handlers for contribution association prompt opt-out
Add a parameter to RegisterForEventHandler to let users set this option.
As noted in the code, this also works for events that have ended or do
not track contributions, which in the UI would not show the checkbox at
all. This is harmless, and it could be fixed later on if needed (we'd
have to add a bit of ad-hoc validation and error messages). Also, note
that this parameter is required and does not provide a default,
consistently with the fact that PUT handlers require a full
representation of the resource being modified, and we wouldn't want a
client to unintentionally use the default value.

Update GetOwnRegistrationInfoHandler to expose the value in the
response, so that the two endpoints can still be used together.

Bug: T408460
Change-Id: Id4e6bdb7601cda40ebfc0e8f1ccf2bf61eb72046