SpecialRegisterForEvent: simplify opt-out handling
Change the conditional for showing the field to check `isEdit`, so the
field can also appear immediately after the first submission, without
having to refresh first.

Use a hidden form field instead of omitting it altogether when the user
shouldn't see anything. This way we can simplify the submission handler.
In fact, it previously wasn't working because HMTLForm automatically
fills in the submission value with `false` after the first form
submission, due to the value not having been submitted.

This addresses the two issues described in T408460#11478800.

Bug: T408460
Change-Id: I22b718278cbe5eab681a27b703bd58d74068cafb